requirements.txt
# Ad AI Auto — Dependencies
google-ads>=24.0.0
facebook-business>=19.0.0
anthropic>=0.40.0
openai>=1.50.0
python-dotenv>=1.0.0
